update faq/nav and readme files
