+++ README.md
@@ -1,15 +1,21 @@
-# Thinking Beyond Rich User Interfaces - Senior Project 2024
+# Thinking Beyond Rich User Interfaces - Senior Project 2024-2025
 
 ## Overview
 
 This project focuses on replacing traditional financial platforms that have rich, detailed user interfaces with a Google-like search interface/conversational AI tool.
-The goal is to allow clients to ask questions and receive answers directly from their financial data. This approach aims to improve user experience and reduce dependency on live support agents.
-
+The goal is to allow clients to ask questions and receive answers based on any FAQs, navigation on their account, or about their financial data. This approach aims to improve user experience and reduce dependency on live support agents.
+ 
 ## Features
 
 -   **Conversational Interface**: Clients can ask questions in natural language and receive clear, detailed answers.
 -   **Personalized Insights**: Provides tailored answers based on user information.
 -   **Efficiency**: Reduces time spent navigating websites by focusing on delivering answers quickly.
+
+## Agents
+
+-   **FAQ and Navigation Assistant**: handle queries related to where services are located or finace related questions.
+-   **User Retrieval**: retrieve user financial data, such as martial status or portfolio numbers, etc.
+-   **Services**: handles queries related to services or the services the user has.
 
 ## Setup
